+++ trunk/stdnet/exceptions.py
@@ -2,6 +2,9 @@
 class StdNetException(Exception):
     pass
 
+class ModelNotRegistered(StdNetException):
+    pass
+
 class ImproperlyConfigured(StdNetException):
     "stdnet is somehow improperly configured"
     pass