+++ trunk/CHANGELOG.rst
@@ -1,7 +1,10 @@
 
 python-stdnet 0.2.3
 ========================================
- * StdModel raise AttributEerror when method/attribute not available. Previously it return None.
+ * Added ListField with Redis implementation.
+ * StdModel raise AttributError when method/attribute not available. Previously it returned None.
+ * StdModel raise ModelNotRegistered when trying to save an instance of a non-registered model.
+ * 13 tests.
 
 python-stdnet 0.2.2 - 2010 July 7
 ========================================